Adding sensor sync and servo related files

+++ src/discoverylab/telebot/master/arms/parser/YEIDataParser.java
@@ -12,9 +12,9 @@ public class YEIDataParser extends CoreDataParser{
 		StringTokenizer tokenizer = parseUsingTokenizer(str);
 		YEIDataModel instance = new YEIDataModel();
 		instance.setJointType(tokenizer.nextToken());
-		instance.setX(Integer.parseInt(tokenizer.nextToken()));
-		instance.setY(Integer.parseInt(tokenizer.nextToken()));
-		instance.setZ(Integer.parseInt(tokenizer.nextToken()));
+		instance.setX(Long.parseLong(tokenizer.nextToken()));
+		instance.setY(Long.parseLong(tokenizer.nextToken()));
+		instance.setZ(Long.parseLong(tokenizer.nextToken()));
 		return instance;
 	}
 }
